lib/AI Models/AI Chatbot Seasonal Advisior/requirements.txt
fastapi
uvicorn
langchain-google-genai
faiss-cpu
chromadb
unstructured
pypdf
tiktoken
python-dotenv
streamlit
langchain
datetime
langchain_community
